requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
sentence-transformers
streamlit
python-dotenv
